requirements.txt
aiogram==3.12.0
gspread==6.0.0
google-auth==2.23.4
requests==2.31.0
icalendar==5.0.11
fastapi==0.104.1
uvicorn==0.24.0
aioschedule==0.4.0
python-dotenv==1.0.0
pandas==2.1.4
